+++ app.js
@@ -1,43 +1,9 @@
 const http = require('http');
-const fs = require('fs');
 
-const server = http.createServer((req, res) => {
-    const url = req.url;
-    const method = req.method;
-    
-    if (url === '/') {
-        res.write('<html>');
-        res.write('<head><title>Enter Message</title><head>');
-        res.write('<body><form action="/message" method="POST"><input type="text" name="message"><button type="submit">Send</button></form></body>');
-        res.write('</html>');
-        return res.end();
-    }
+const routes = require('./routes');
 
-    if (url === '/message' && method === 'POST') {
-        const body = [];
-        req.on('data', (chunk) => {
-            console.log(chunk);
-            body.push(chunk);
-        });
-        return req.on('end', () => {
-            const parsedBody = Buffer.concat(body).toString();
-            const message = parsedBody.split('=')[1];
-            // This make the code stop and only continue when the file is finished.
-            // fs.writeFileSync('message.txt', message);
-            fs.writeFile('message.txt', message, error => {
-                res.statusCode = 302;
-                res.setHeader('Location', '/');
-                return res.end();
-            });
-        });
-    }
+console.log(routes.someText);
 
-    res.setHeader('Content-Type', 'text/html');
-    res.write('<html>');
-    res.write('<head><title>My First Page</title><head>');
-    res.write('<body><h1>Hello from my Node.js Server!</h1></body>');
-    res.write('</html>');
-    res.end();
-});
+const server = http.createServer(routes.handler);
 
 server.listen(3000);
