+++ routes.js
@@ -1,0 +1,54 @@
+const fs = require('fs');
+
+const requestHandler = (req, res) => {
+    const url = req.url;
+    const method = req.method;
+
+    if (url === '/') {
+        res.write('<html>');
+        res.write('<head><title>Enter Message</title><head>');
+        res.write('<body><form action="/message" method="POST"><input type="text" name="message"><button type="submit">Send</button></form></body>');
+        res.write('</html>');
+        return res.end();
+    }
+    
+    if (url === '/message' && method === 'POST') {
+        const body = [];
+        req.on('data', (chunk) => {
+            console.log(chunk);
+            body.push(chunk);
+        });
+        return req.on('end', () => {
+            const parsedBody = Buffer.concat(body).toString();
+            const message = parsedBody.split('=')[1];
+            // This make the code stop and only continue when the file is finished.
+            // fs.writeFileSync('message.txt', message);
+            fs.writeFile('message.txt', message, error => {
+                res.statusCode = 302;
+                res.setHeader('Location', '/');
+                return res.end();
+            });
+        });
+    }
+    
+    res.setHeader('Content-Type', 'text/html');
+    res.write('<html>');
+    res.write('<head><title>My First Page</title><head>');
+    res.write('<body><h1>Hello from my Node.js Server!</h1></body>');
+    res.write('</html>');
+    res.end();
+};
+
+// Line by line export
+// module.exports.handler = requestHandler;
+// module.exports.someText = 'Some hard coded text';
+
+// Line by line export short syntax
+// exports.handler = requestHandler;
+// exports.someText = 'Some hard coded text';
+
+// Object export
+module.exports = {
+    handler: requestHandler,
+    someText: 'Some hard coded text'
+};
